Changes to integer division warning implementation.

+++ ninja_ide/gui/menus/menubar.py
@@ -97,7 +97,7 @@ class _MenuBar(QObject):
                     and ((namespace == child_namespace) or (not namespace)):
                 child, weight = self._children[each_child]
                 #Group by namespace and weight
-                weight_index = "%d_%s" % (weight / 100.0, namespace)
+                weight_index = "%d_%s" % (int(weight / 100.0), namespace)
                 children[weight_index].append((child, weight))
 
         return children
